renamed question to summary-prompt
